indexer/requirements.txt
llama-index-embeddings-huggingface
llama-index-core
llama-index-readers-mongodb
llama-index-vector-stores-chroma
pymongo
chromadb 